upload-service/requirements.txt
fastapi
uvicorn

python-multipart
kafka-python